add "enumflags" feature which uses the enum discriminators directly
Some use FFI use cases in C twist the notion of enum
to mean "a symbol holding some value". Thus, they set the
value of the C enum to be a direct bit value.

This patch is a prototype to allow to replicate such a functionality
by enabling the "enumflags" feature.